+++ src/lib.rs
@@ -51,6 +51,7 @@ extern crate log;
 #[macro_use]
 extern crate error_chain;
 extern crate raii_change_tracker;
+#[cfg(feature = "bundle_files")]
 extern crate includedir;
 extern crate serde;
 extern crate serde_json;